chore: validate ở role và xử lý đầu vào ở status

+++ ec-project-api/Services/roles/RoleService.cs
@@ -9,13 +9,16 @@ namespace ec_project_api.Services
     public class RoleService : BaseService<Role, short>, IRoleService
     {
         private readonly IRolePermissionRepository _rolePermissionRepository;
+        private readonly IUserRepository _userRepository;
 
         public RoleService(
             IRoleRepository roleRepository,
-            IRolePermissionRepository rolePermissionRepository
+            IRolePermissionRepository rolePermissionRepository,
+            IUserRepository userRepository
         ) : base(roleRepository)
         {
             _rolePermissionRepository = rolePermissionRepository;
+            _userRepository = userRepository;
         }
 
         public async Task<bool> AssignPermissionsAsync(short roleId, IEnumerable<short> permissionIds)
@@ -60,7 +63,7 @@ namespace ec_project_api.Services
 
         public override async Task<Role?> GetByIdAsync(short id, QueryOptions<Role>? options = null)
         {
-            options = new QueryOptions<Role>();
+            options ??= new QueryOptions<Role>();
             options.Includes.Add(r => r.Status);
             options.Includes.Add(r => r.RolePermissions);
 
@@ -77,6 +80,13 @@ namespace ec_project_api.Services
             var role = await _repository.GetByIdAsync(id);
             if (role == null)
                 throw new KeyNotFoundException(RoleMessages.RoleNotFound);
+
+            var usersWithRole = await _userRepository.FindAsync(
+                u => u.UserRoleDetails.Any(urd => urd.RoleId == id)
+            );
+
+            if (usersWithRole.Any())
+                throw new InvalidOperationException(RoleMessages.UserAssignedRole);
 
             await _repository.DeleteAsync(role);
             await _repository.SaveChangesAsync();
